Fix res issue, add create user, add list profile, add a used column in db

+++ src/view/product/delete.product.ts
@@ -1,6 +1,7 @@
 import { requireUser } from "../../middleware/requireUser.js";
 import { express } from "../../config/common.js";
 import { deleteProductHandler } from "../../controllers/product/deleteproduct.controler.js";
+import { handelErrorResponse } from "../../controllers/response.controller.js";
 
 const router = express.Router();
 const app = express();
@@ -9,8 +10,16 @@ app.use(express.urlencoded({ extended: true }));
 
 router.delete("/", (req, res) => {
     requireUser(req, res, async ({ id, dbCode }: { id: number, dbCode: string }) => {
-        const productId = req.body.id
-        await deleteProductHandler(productId, id, dbCode, res)
+        try {
+            const productId = req.body.id
+            if (!productId) {
+                handelErrorResponse(res, "Product Id is required")
+            } else {
+                await deleteProductHandler(productId, id, dbCode, res)
+            }
+        } catch (err) {
+            await handelErrorResponse(res, "Something went wrong")
+        }
     })
 });
